load && 50 not in 100 set via rcsb?  Changes to be committed: 	modified:   build_cys-sqlite.log 	new file:   db_loads/2019-04-10_loadlist_0.json 	modified:   script/Perl/update_scripts/util/populate_exp_method_identity.pl
For first time, the script/Perl/update_scripts/util/populate_exp_method_identity.pl
was throwing an exception based on comparisons between the 50 and 100%
identity sets.  All the 50% pdb_ids are expected to exist in the 100% set, and that
has continued to be true for some time.  This week, 50% entries have
been identified to be missing in the 100% set. Something may have
changed in RCSB services.

script/Perl/update_scripts/util/populate_exp_method_identity.pl now
dumps those observed:

$VAR1 = {
          '50 not in 100' => {
                               'SOLUTION NMR' => [
                                                   '2n8c'
                                                 ],
                               'X-RAY DIFFRACTION' => [
                                                        '1ldj',
                                                        '1sl4',
                                                        '3uf2',
                                                        '4ia0',
                                                        '4xiz',
                                                        '5urv'
                                                      ]
                             }
        };

along with the xml queries along with a warning rather than a die.